requirements.txt
torch
torchvision
deepspeed
toml
transformers
diffusers>=0.32.1
datasets
pillow
sentencepiece
protobuf
peft
torch-optimi
tensorboard
tqdm
safetensors
bitsandbytes
imageio[ffmpeg]
av
einops
accelerate
loguru
flash-attn
omegaconf
iopath
termcolor
hydra-core
easydict
ftfy
gradio
wandb
#sageattention==1.0.6